% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_utilities.R
\name{anova.svcm}
\alias{anova.svcm}
\title{Deviance tables.}
\usage{
\method{anova}{svcm}(object, ...)
}
\arguments{
\item{object}{An object of type \code{svcm}.}

\item{...}{list of \code{svcm} models.}
}
\value{
Anova.
}
\description{
Produce deviance tables for \code{svcm} objects.
}
