% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_utilities.R
\name{coef.fitm}
\alias{coef.fitm}
\title{Returns fitted model parameters.}
\usage{
\method{coef}{fitm}(object, ...)
}
\arguments{
\item{object}{An object of type fitm.}

\item{...}{Not used.}
}
\value{
Fitted parameters.
}
\description{
Returns fitted model parameters for fitm objects.
}
