% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_utilities.R
\name{coef.summary.fitm}
\alias{coef.summary.fitm}
\title{Returns fitted model parameters.}
\usage{
\method{coef}{summary.fitm}(object, ...)
}
\arguments{
\item{object}{An object of class summary.fitm.}

\item{...}{Not used.}
}
\value{
\code{data.frame} of model parameters with details.
}
\description{
Returns detailed table of model parameters for objects of class \code{summary.fitm}.
}
