% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{compHess2}
\alias{compHess2}
\title{Compute hessian}
\usage{
compHess2(mod, ...)
}
\arguments{
\item{mod}{An object of type \code{fitm}}

\item{...}{Arguments passed to \code{numDeriv::hessian}.}
}
\value{
Hessian matrix.
}
\description{
Computes hessian
}
