% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_utilities.R
\name{compute.fitm}
\alias{compute.fitm}
\title{Computes functions of parameters in fitted model objects.}
\usage{
\method{compute}{fitm}(object, form, ...)
}
\arguments{
\item{object}{A Fitted model of type fitm.}

\item{form}{An expression to be evaluated.}

\item{...}{Not used.}
}
\description{
Can be used to evaluate expressions containing functions of parameters.
}
