% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_utilities.R
\name{compute}
\alias{compute}
\title{Computes functions of parameters.}
\usage{
compute(object, ...)
}
\arguments{
\item{object}{Either a model og type svcm or a fitted model of type fitm.}

\item{...}{Further function arguments.}
}
\description{
Can be used to evaluate expressions containing functions of parameters.
}
