% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_utilities.R
\name{compute.svcm}
\alias{compute.svcm}
\title{Computes functions of parameters in model objects.}
\usage{
\method{compute}{svcm}(object, form, ...)
}
\arguments{
\item{object}{A model of type svcm.}

\item{form}{An expression to be evaluated.}

\item{...}{Not used.}
}
\description{
Can be used to evaluate expressions containing functions of model objects.
}
