% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{const}
\alias{const}
\title{Constructor for constant object}
\usage{
const(value, name)
}
\arguments{
\item{value}{The constant value.}

\item{name}{Character giving name to reference the constant in model expressions.}
}
\value{
An object of class \code{const}.
}
\description{
Creates a constant object for use in model expressions. Constants
are values (e.g. sparse matrices) that are referenced in \code{svc}, \code{mc},
or \code{ic} expressions but are not model parameters.
}
