% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{dat_svcm}
\alias{dat_svcm}
\title{Create a object for storing data}
\usage{
dat_svcm(Y)
}
\arguments{
\item{Y}{response data described by model.}
}
\value{
an object of class \code{dat_svcm}.
}
\description{
creates a \code{dat_svcm} object for the response variable.
}
