% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{datm}
\alias{datm}
\title{Create a object for storing data}
\usage{
datm(Y)
}
\arguments{
\item{Y}{data described by model.}
}
\value{
an object of class \code{datm}.
}
\description{
creates a \code{datm} object.
}
