% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{.compute.fixedmc}
\alias{.compute.fixedmc}
\title{Compute function for mc object}
\usage{
\method{.compute}{fixedmc}(object, env, ...)
}
\arguments{
\item{object}{An object of type \code{mc}.}

\item{env}{Computing environment.}

\item{...}{Not used.}
}
\value{
vector/matrix with means for an object of type \code{mc}.
}
\description{
Internal function used to evaluate expressions containing objects of type \code{fixedmc}.
}
