% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{.compute.free}
\alias{.compute.free}
\title{Compute function for svc object.}
\usage{
\method{.compute}{free}(object, env, ...)
}
\arguments{
\item{object}{An object of type \code{free}.}

\item{env}{Computing environment.}

\item{...}{Not used.}
}
\description{
Internal function used to evaluate expressions containing objects of type \code{free}.
}
