% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{.computeC.fixedmc}
\alias{.computeC.fixedmc}
\title{Compute function for mc object}
\usage{
\method{.computeC}{fixedmc}(object, env, ...)
}
\arguments{
\item{object}{An object of type mc.}

\item{env}{Computing environment.}

\item{...}{Not used.}
}
\description{
Internal function used to evaluate expressions containing objects of type \code{fixedmc}.
}
