% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{.computeC.fixedsvc}
\alias{.computeC.fixedsvc}
\title{Compute function for svc objects}
\usage{
\method{.computeC}{fixedsvc}(object, env, ...)
}
\arguments{
\item{object}{An object of type svc.}

\item{env}{Computing environment.}

\item{...}{Not used.}
}
\description{
Internal function used to evaluate expressions containing objects of type \code{fixedsvc}.
}
