% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{.computeC}
\alias{.computeC}
\title{Generic compute function.}
\usage{
.computeC(object, ...)
}
\arguments{
\item{object}{An object of type mc/svc.}

\item{...}{Further function arguments.}
}
\description{
Internal functions used to evaluate expressions containing functions of model objects.
}
