% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{.computeC.svc}
\alias{.computeC.svc}
\title{Compute function for svc object.}
\usage{
\method{.computeC}{svc}(object, env, ...)
}
\arguments{
\item{object}{An object of type svc.}

\item{env}{Computing environment.}

\item{...}{Not used.}
}
\description{
Internal function used to evaluate expressions containing functions of model objects.
}
