% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{.computeCMissy.mc}
\alias{.computeCMissy.mc}
\title{Compute function for mc object.}
\usage{
\method{.computeCMissy}{mc}(object, env, missy, ...)
}
\arguments{
\item{object}{An object of type mc.}

\item{env}{Computing environment.}

\item{...}{Not used.}
}
\description{
Internal function used to evaluate expressions containing functions of model objects.
}
