% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{.get_free_labels}
\alias{.get_free_labels}
\title{Labels of free values in parameter matrix}
\usage{
.get_free_labels(pm)
}
\arguments{
\item{pm}{an instance of an \code{pm} object.}
}
\value{
vector of labels of free elements in parameter matrix (column wise).
}
\description{
Labels of free values in parameter matrix
}
