% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{.get_free_values}
\alias{.get_free_values}
\title{Free values in parameter matrix}
\usage{
.get_free_values(pm)
}
\arguments{
\item{pm}{an instance of an \code{pm} object.}
}
\value{
vector of values of free elements in parameter matrix (column wise)
}
\description{
Free values in parameter matrix
}
