% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{.ldmvnorm}
\alias{.ldmvnorm}
\title{Multivariate normal log-density. Copied from https://github.com/cran/sparseMVN}
\usage{
.ldmvnorm(x, mu, CH)
}
\arguments{
\item{x}{Data vector.}

\item{mu}{Mean vector.}

\item{CH}{Result from Matrix::Cholesky}
}
\value{
Log likelihood.
}
\description{
Computes multivariate normal log-density based on sparse matrix types.
}
