% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{.objective}
\alias{.objective}
\title{Objective function}
\usage{
.objective(y, M, S)
}
\arguments{
\item{y}{Data vector.}

\item{M}{Mean vector.}

\item{S}{Covariance matrix.}
}
\value{
Twice negative log likelihood.
}
\description{
Objetive function form multivariate normal
}
