% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{.update_ic}
\alias{.update_ic}
\title{Updates intermediate computations}
\usage{
.update_ic(ic, env_comp)
}
\arguments{
\item{env_comp}{computing environment where the expression is evaluated and stored.}

\item{pm}{parameter matrix object.}
}
\description{
Used during optimization for computing expressions in \code{ic} objects.
}
