% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{.update_ics}
\alias{.update_ics}
\title{Updates all intermediate computation objects of a model}
\usage{
.update_ics(mod)
}
\arguments{
\item{mod}{an instance of an \code{svcm} model.}
}
\description{
Used during optimization for updating intermediate computations in computing environment.
}
