% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{.update_pm}
\alias{.update_pm}
\title{Updates values in parameter matrix}
\usage{
.update_pm(pm, theta, env_comp)
}
\arguments{
\item{pm}{an instance of an \code{pm} object.}

\item{theta}{vector of free model parameters.}

\item{env_comp}{computing environment where the expression is evaluated and stored.}
}
\description{
Used during optimization for updating free values in a parameter matrix in both computing environment and model object.
}
