% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{.update_pms}
\alias{.update_pms}
\title{Updates all \code{pm} objects of a model}
\usage{
.update_pms(m, theta)
}
\arguments{
\item{m}{an instance of an \code{svcm} model.}

\item{theta}{vector of free model parameters.}
}
\description{
Used during optimization for updating parameter matrices in computing environment and model object.
}
