% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{.updateValues}
\alias{.updateValues}
\title{Updates model objects}
\usage{
.updateValues(mo, values, env_comp)
}
\arguments{
\item{mo}{model object.}

\item{values}{free parameters.}

\item{env_comp}{computing environment.}
}
\description{
Used during optimization for updating model objects in computing environment.
}
