% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{expected_cov}
\alias{expected_cov}
\title{Model implied covariance matrix}
\usage{
expected_cov(svcm, drop_miss = T)
}
\arguments{
\item{svcm}{an instance of an \code{svcm} model.}

\item{drop_miss}{drop expectation for missing values?}
}
\value{
matrix of (co)variances.
}
\description{
Compute model implied covariance matrix.
}
