% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{expected_mean}
\alias{expected_mean}
\title{Model implied mean vector}
\usage{
expected_mean(mod, drop_miss = T)
}
\arguments{
\item{mod}{an instance of an \code{svcm} model.}

\item{drop_miss}{drop expectation for missing values?}
}
\value{
vector of means.
}
\description{
Compute model implied mean vector.
}
