% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{expectedM}
\alias{expectedM}
\title{Model implied mean vector}
\usage{
expectedM(svcm, dropmiss = T)
}
\arguments{
\item{svcm}{an instance of an \code{svcm} model.}

\item{dropmiss}{drop expectation for missing values?}
}
\value{
vector of means.
}
\description{
Compute model implied mean vector.
}
