% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{expectedS}
\alias{expectedS}
\title{Model implied covariance matrix}
\usage{
expectedS(svcm, dropmiss = T)
}
\arguments{
\item{svcm}{an instance of an \code{svcm} model.}

\item{dropmiss}{drop expectation for missing values?}
}
\value{
matrix of (co)variances.
}
\description{
Compute model implied covariance matrix.
}
