% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{fd_hess_svcm}
\alias{fd_hess_svcm}
\title{Compute hessian}
\usage{
fd_hess_svcm(mod, ...)
}
\arguments{
\item{mod}{An object of type \code{fitm}}

\item{...}{Arguments passed to \code{numDeriv::hessian}.}
}
\value{
Hessian matrix.
}
\description{
Computes hessian of parameters with finite differences
}
