% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{fd_jacobian_svcm}
\alias{fd_jacobian_svcm}
\title{Compute jacobian of expression}
\usage{
fd_jacobian_svcm(mod, form, ...)
}
\arguments{
\item{mod}{An object of type \code{svcm}}

\item{form}{An expression for the calculation.}

\item{...}{Arguments passed to \code{numDeriv::hessian}.}
}
\value{
Jacobian matrix.
}
\description{
Computes jacobian of expression containing parameter matrices with finite differences. The output should be a vector.
Note that the jacobian is computed with regard to all free parameters in the model.
}
