% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVoxelSequenceRealdata.R
\name{findVoxelSequenceRealdata}
\alias{findVoxelSequenceRealdata}
\title{Sort voxel sequence in distant order for real data}
\usage{
findVoxelSequenceRealdata(wmNonzeroId,dimn)
}
\arguments{
\item{wmNonzeroId}{the index for nonzerod Id}
\item{dimn}{the dimension of image data}
}
\description{
Sort voxel sequence in distant order for real data
}
\examples{
  wmNonzeroId=1:213037
  dimn=c(128,128,96)
  a=findVoxelSequenceRealdata(wmNonzeroId,dimn);
  vxlNbySeqid=a[[1]];
  vxlDistSeq=a[[2]];
  vxlNberSeq=a[[3]];

}

