% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVoxelSequenceRealdata_para.R
\name{findVoxelSequenceRealdata_para}
\alias{findVoxelSequenceRealdata_para}
\title{Sort voxel sequence in distant order for real data using parallel computing}
\usage{
findVoxelSequenceRealdata_para(wmNonzeroId,dimn,cores=2)
}
\arguments{
\item{wmNonzeroId}{the index for nonzerod Id}
\item{dimn}{the dimension of image data}
\item{cores}{number of cores used in parallel computing}
}
\description{
Sort voxel sequence in distant order for real data using parallel computing
}
\examples{
  wmNonzeroId=1:213037
  dimn=c(128,128,96)
  a=findVoxelSequenceRealdata_para(wmNonzeroId,dimn,cores=2);
  vxlNbySeqid=a[[1]];
  vxlDistSeq=a[[2]];
  vxlNberSeq=a[[3]];

}

