% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{fit_svcm}
\alias{fit_svcm}
\title{Fit a model}
\usage{
fit_svcm(mod, se = FALSE, ...)
}
\arguments{
\item{mod}{an instance of an \code{svcm} model.}

\item{se}{should standard errors be computed?}

\item{...}{arguments passed to \code{nlminb}.}
}
\value{
an object of class \code{svcm}.
}
\description{
fits a \code{svcm} model.
}
