% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{fitm}
\alias{fitm}
\title{Fit a model}
\usage{
fitm(svcm, se = FALSE, ...)
}
\arguments{
\item{svcm}{an object of class \code{svcm}.}

\item{se}{dhould standard errors be computed?}

\item{...}{arguments passed to \code{nlminb}.}
}
\value{
an object of class \code{svcm}.
}
\description{
fits a \code{svcm} model.
}
