% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_utilities.R
\name{logLik.fitm}
\alias{logLik.fitm}
\title{Computes log-likelihood for fitm objects.}
\usage{
\method{logLik}{fitm}(object, ...)
}
\arguments{
\item{object}{An object of type fitm.}

\item{...}{Not used.}
}
\description{
Computes log-likelihood for fitm objects.
}
