% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{mc}
\alias{mc}
\title{Constructor for mean component object}
\usage{
mc(form, X = NULL)
}
\arguments{
\item{form}{An expression for the mean model.}

\item{X}{A design matrix.}
}
\value{
An object of class \code{mc}.
}
\description{
Creates a mean component object used to define (part of) the mean structure in the model.
}
\examples{
library(svcmr)
X <- cbind(1, rnorm(100))
B <- pm(nrow = 6, ncol = 2,
        labels = paste0("b", 1:12),
        values = 0,
        free = TRUE,
        name = "B")
mc <- mc(form = B, X = X)
}
