% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrWresd.cpp
\name{mlrWresd}
\alias{mlrWresd}
\title{Get the least squares estimates of coefficient functions }
\usage{
mlrWresd(X,Y)
}
\arguments{
\item{X}{design matrix}

\item{Y}{image data}

}
\description{
Get the least squares estimates of ¦Â 
}
\examples{
  n=60;
  xMatrix =matrix(rep(1,n*3),n);
  xMatrix[,2]=rnorm(n,1,0.5);
  xMatrix[,3]=runif(n,1,2);
  xMatrix[,2] = (xMatrix[,2]-mean(xMatrix[,2]))/sd(xMatrix[,2]);
  xMatrix[,3] = (xMatrix[,3]-mean(xMatrix[,3]))/sd(xMatrix[,3]);
  imgData = matrix(rnorm(n*32768),ncol=64*64*8);
  s=mlrWresd(xMatrix,imgData);
  mxBeta=s[[1]];
  mxCovb=s[[2]];
  mxR=s[[3]];
}

