% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_utilities.R
\name{nobs.fitm}
\alias{nobs.fitm}
\title{Returns number of observations used to fit model.}
\usage{
\method{nobs}{fitm}(object, ...)
}
\arguments{
\item{object}{An object of type fitm.}

\item{...}{Not used.}
}
\value{
Number of observations.
}
\description{
Returns number of observations used to fit model.
}
