% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{objective}
\alias{objective}
\title{Objective function}
\usage{
objective(mod)
}
\arguments{
\item{mod}{An object of type \code{fitm}}
}
\value{
Twice negative log likelihood.
}
\description{
Objective function
}
