% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{pm}
\alias{pm}
\title{Constructor for parameter matrix}
\usage{
pm(
  nrow = NA,
  ncol = NA,
  labels = character(),
  free = FALSE,
  values = NA,
  name = NA
)
}
\arguments{
\item{nrow}{Number of matrix rows.}

\item{ncol}{Number of matrix columns.}

\item{labels}{Vector/matrix with parameter labels. Equal characters can be used to define equality constraints.}

\item{free}{Vector or matrix of logical values defining whether parameters are free or not.}

\item{values}{Vector/matrix with parameter values. Values for free parameters are used as starting values during optimization
and values for fixed parameters are constants during optimization.}

\item{name}{Character giving name to model object.}
}
\value{
An object of class \code{pm}.
}
\description{
Creates an object of type \code{pm} used to represent parameters in a model.
}
\examples{
library(svcmr)
S <- pm(nrow = 2, ncol = 2,
        labels = c("s11", "s12", "s12", "s22"),
        values = c(2, 1, 1, 2),
        free = TRUE,
        name = "S")
}
