% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_print.R
\name{print.fe}
\alias{print.fe}
\title{Print function for fe objects.}
\usage{
\method{print}{fe}(x, ...)
}
\arguments{
\item{x}{An object of type fe.}

\item{...}{Not used}
}
\description{
Print function for fe (profiled fixed-effect) objects.
}
