% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_print.R
\name{print.mc}
\alias{print.mc}
\title{Print function for mc objects.}
\usage{
\method{print}{mc}(x, ...)
}
\arguments{
\item{x}{An object of type mc.}

\item{...}{Not used}
}
\description{
Print function for mc objects.
}
