% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_utilities.R
\name{SEm}
\alias{SEm}
\title{Compute hessian}
\usage{
SEm(fitm)
}
\arguments{
\item{object}{A fitted model of type \code{fitm}.}
}
\description{
Computes hessian for fitted models of type \code{fitm}.
}
