% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simtest.R
\name{Simtest}
\alias{Simtest}
\title{Do simulation studies}
\usage{
Simtest(n, pattern, pr, nf1,ef, f1,f2)
}
\arguments{
\item{n}{the number of subject}

\item{pattern}{There are seven patterns to generate parameter slice}

\item{pr}{a parameter that we need to adjust in simulation study}

\item{nf1}{a parameter that we need to adjust in simulation study}

\item{ef}{It decides the distribution of noises error we use in simulation study. 1 means normal distribution and 2 means chi-square distribution}

\item{f1}{a parameter that we need to adjust in simulation study}

\item{f2}{a parameter that we need to adjust in simulation study}
}
\description{
Do simulation studies
}
\examples{
  pattern=6;
  pr=0.5;
  nf1=ef=f1=f2=1;
  n=60;
  Simtest(n, pattern, pr, nf1,ef, f1,f2);

}

