% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothAspResidualRealdata.R
\name{smoothAspResidualRealdata}
\alias{smoothAspResidualRealdata}
\title{Calculate and smooth residuals after adaptive smoothing parameter and do the Eigen-decomposition }
\usage{
smoothAspResidualRealdata(Xdesign,Imgdata,aspBeta,wmNonzeroId,dimn,part)
}

\arguments{
\item{Xdesign}{the design matrix}

\item{Imgdata}{the image data}

\item{aspBeta}{smoothed beta}

\item{wmNonzeroId}{the index of nonzero id}

\item{dimn}{the dimension of image data}

\item{part}{the dimension of the unit after parting the entire image}

}

\description{
Calculate and smooth residuals after adaptive smoothing parameter
}

\examples{
  Xdesign=matrix(rnorm(174*8),nrow=174)
  Imgdata=matrix(rnorm(174*213037),nrow=174)
  aspBeta=matrix(rnorm(8*213037),nrow=8)
  wmNonzeroId=1:213037;
  dimn=c(128,128,96);
  part=c(8,8,6);
  a=smoothAspResidualRealdata(Xdesign,Imgdata,aspBeta,wmNonzeroId,dimn,part);
}

