% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothParameterRealdata.R
\name{smoothParameterRealdata}
\alias{smoothParameterRealdata}
\title{Use MASS method to individually smooth each least squares estimate for real data}
\usage{
smoothParameterRealdata(mxBeta,covEvector,vxlNberSeq,vxlNbySeqid,vxlDistSeq,xMatrix,covEvalue,sigmaError)
}

\arguments{
\item{vxlNbySeqid}{the id of nearby voxel}

\item{vxlDistSeq}{the distance of nearby voxel sequences}

\item{vxlNberSeq}{the maximum number of nearby voxel}

\item{covEvector}{the eigen functions of the covariance matrix of ¦Ç}

\item{covEvalue}{the eigen values of the covariance matrix of ¦Ç}

\item{mxBeta}{the beta matrix from least square estimates}

\item{xMatrix}{the design matrix}

\item{sigmaError}{the variance vector of measurement error}

}

\description{
Use MASS method to individually smooth each least squares estimate for real data
}

\examples{
  see the example in SVCM package
}

