% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothResidualRealdata.R
\name{smoothResidualRealdata}
\alias{smoothResidualRealdata}
\title{Smooth the individual residuals}
\usage{
smoothResidualRealdata(Xdesign,ImgData,wmNonzeroId,dim,part)
}
\arguments{
\item{Xdesign}{The design matrix}

\item{ImgData}{The image data}

\item{wmNonzeroId}{The index of nonzero Id}

\item{dim}{The dimension of the image data}

\item{part}{The dimension of the unit after parting the entire image}

}
\description{
Smooth the individual residuals 
}
\examples{
 n=174;
 wmNonzeroId=1:213037
 ImgData=matrix(rnorm(213037*n),nrow=n)
 Xdesign=matrix(rnorm(8*n),nrow=n)
 a1=c(128,128,96)
 a2=c(8,8,6)
 a=smoothResidualRealdata(Xdesign,ImgData,wmNonzeroId,a1,a2);
 mxBeta=a[[1]];
 sigmaError=a[[2]];
 covEvalue=a[[3]];
 covEvector=a[[4]];
}

