% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothResidualRealdata_para.R
\name{smoothResidualRealdata_para}
\alias{smoothResidualRealdata_para}
\title{ Smooth the individual residuals using parallel computing using parallel computing}
\usage{
smoothResidualRealdata_para(Xdesign,ImgData,wmNonzeroId,dim,part,cores=2)
}
\arguments{
\item{Xdesign}{The design matrix}

\item{ImgData}{The image data}

\item{wmNonzeroId}{The index of nonzero Id}

\item{dim}{The dimension of the image data}

\item{part}{The dimension of the unit after parting the entire image}

\item{cores}{number of cores used in parallel computing}


}
\description{
Smooth the individual residuals using parallel computing
}
\examples{
 n=174;
 wmNonzeroId=1:213037
 ImgData=matrix(rnorm(213037*n),nrow=n)
 Xdesign=matrix(rnorm(8*n),nrow=n)
 a1=c(128,128,96)
 a2=c(8,8,6)
 a=smoothResidualRealdata_para(Xdesign,ImgData,wmNonzeroId,a1,a2,cores=2);
 mxBeta=a[[1]];
 sigmaError=a[[2]];
 covEvalue=a[[3]];
 covEvector=a[[4]];
}

