% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_utilities.R
\name{summary.fitm}
\alias{summary.fitm}
\title{Summary function for fitm objects.}
\usage{
\method{summary}{fitm}(object, ...)
}
\arguments{
\item{object}{An object of type fitm.}

\item{...}{Not used.}
}
\description{
Summary function for fitm objects.
}
