% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_utilities.R
\name{summary.svcm}
\alias{summary.svcm}
\title{Summary function for svcmr models.}
\usage{
\method{summary}{svcm}(object, ...)
}
\arguments{
\item{object}{An object of type \code{fitm}.}

\item{...}{Not used.}
}
\description{
Summary function for class \code{"fitm"}.
}
