% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\docType{package}
\name{svcm-package}
\alias{svcm-package}
\title{svcm: Structured Variance Component Models}
\description{
Fit variance component models with sparse relationship matrices.
}
\author{
\strong{Maintainer}: Espen Moen Eilertsen \email{e.m.eilertsen@psykologi.uio.no} (\href{https://orcid.org/0000-0003-3444-4251}{ORCID})

Authors:
\itemize{
  \item Espen Moen Eilertsen \email{e.m.eilertsen@psykologi.uio.no} (\href{https://orcid.org/0000-0003-3444-4251}{ORCID})
}

}
