% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{theta}
\alias{theta}
\title{Returns parameters of a model}
\usage{
theta(svcm)
}
\arguments{
\item{svcm}{an object of class \code{svcm}.}
}
\value{
vector of parameters.
}
\description{
returns non-duplicated parameters of a \code{svcm} model.
}
