% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{update_model}
\alias{update_model}
\title{Update parameter matrices in environment from theta}
\usage{
update_model(mod, theta)
}
\arguments{
\item{mod}{An object of type \code{fitm}.}

\item{theta}{A vector of parameter values.}
}
\description{
update all free elements in parameter matrices.
}
