% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_utilities.R
\name{vcov.svcm}
\alias{vcov.svcm}
\title{Returns covariance matrix of fitted model parameters.}
\usage{
\method{vcov}{svcm}(object, ...)
}
\arguments{
\item{object}{An object of type fitm.}

\item{...}{Not used.}
}
\value{
Covariance matrix of fitted parameters.
}
\description{
Returns covariance matrix of fitted model parameters for fitm objects.
}
