% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_utilities.R
\name{with.svcm}
\alias{with.svcm}
\title{Evaluate an Expression in an svcm Model Environment}
\usage{
\method{with}{svcm}(data, expr, ...)
}
\arguments{
\item{data}{an instance of an \code{svcm} model.}

\item{expr}{An expression to be evaluated.}

\item{...}{Not used.}
}
\description{
S3 method for \code{with()} to evaluate expressions in the model's
computing environment.
}
