% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_utilities.R
\name{anova.fitm}
\alias{anova.fitm}
\title{Deviance tables.}
\usage{
\method{anova}{fitm}(object, ...)
}
\arguments{
\item{object}{An object of type fitm.}

\item{...}{Not used.}
}
\value{
Anova.
}
\description{
Produce deviance tables for fitm objects.
}
