% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_utilities.R
\name{coef.svcm}
\alias{coef.svcm}
\title{Returns fitted model parameters.}
\usage{
\method{coef}{svcm}(object, ...)
}
\arguments{
\item{object}{An object of type \code{svcm}.}

\item{...}{Not used.}
}
\value{
Vector of fitted parameters.
}
\description{
Returns fitted model parameters for \code{svcm} objects.
}
