% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_utilities.R
\name{compute.fitm}
\alias{compute.fitm}
\title{Computes functions of parameters}
\usage{
\method{compute}{fitm}(object, form, ...)
}
\arguments{
\item{object}{An object of class \code{fitm}.}

\item{form}{An expression to be evaluated.}

\item{...}{Not used.}
}
\description{
Evaluates expressions containing functions of model parameters in objects of class \code{fitm}.
}
