% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_utilities.R
\name{compute}
\alias{compute}
\title{Computes functions of parameters.}
\usage{
compute(object, ...)
}
\arguments{
\item{object}{Either a model of type \code{svcm} or a fitted model of type \code{fitm}.}

\item{...}{Further function arguments.}
}
\description{
\code{compute} is a generic function for evalutatin expressions containing functions of parameters in models.
}
