% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_utilities.R
\name{compute}
\alias{compute}
\title{Computes functions of parameters}
\usage{
compute(mod, form)
}
\arguments{
\item{mod}{an instance of an \code{svcm} model.}

\item{form}{An expression to be evaluated.}
}
\description{
Evaluates expressions containing functions of model parameters in objects of class \code{svcm}.
}
