% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{.compute.fixedsvc}
\alias{.compute.fixedsvc}
\title{Compute function for svc objects}
\usage{
\method{.compute}{fixedsvc}(object, env, ...)
}
\arguments{
\item{object}{An object of type \code{svc}.}

\item{env}{Computing environment.}

\item{...}{Not used.}
}
\value{
Matrix with covariances for an object of type \code{svc}.
}
\description{
Internal function used to evaluate expressions containing objects of type \code{fixedsvc}.
}
