% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{.compute}
\alias{.compute}
\title{Generic compute function}
\usage{
.compute(object, ...)
}
\arguments{
\item{object}{An object of type \code{mc} / \code{svc}.}

\item{...}{Further function arguments.}
}
\description{
Internal functions used to evaluate expressions containing functions of model objects.
}
