% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{.get_free_labels}
\alias{.get_free_labels}
\title{Return labels of free values in parameter matrix as a vector}
\usage{
.get_free_labels(pm)
}
\arguments{
\item{pm}{parameter matrix object.}
}
\description{
Return labels of free values in parameter matrix as a vector
}
