% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_main.R
\name{.get_free_values}
\alias{.get_free_values}
\title{Generic update function
Update function for pm objects
Update function for ic objects
Denne er ikke riktig. Feil signatur. trenger ikke theta (x)
Return free values in parameter matrix as a vector}
\usage{
.get_free_values(pm)
}
\arguments{
\item{pm}{parameter matrix object.}
}
\description{
Generic update function
Update function for pm objects
Update function for ic objects
Denne er ikke riktig. Feil signatur. trenger ikke theta (x)
Return free values in parameter matrix as a vector
}
