% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{fitm}
\alias{fitm}
\title{Fit a model}
\usage{
fitm(svcm, se = FALSE, ...)
}
\arguments{
\item{svcm}{An object of type svcm.}

\item{se}{Should standard errors be computed?}

\item{...}{Arguments passed to nlminb.}
}
\value{
An object of class fitm.
}
\description{
Fits a model returned from svcm.
}
\examples{
library(svcmr)
R <- Matrix::Diagonal(100)
L <- pm(nrow = 6, ncol = 2,
        labels = paste0("l", 1:12),
        values = c(1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1),
        free = c(FALSE, TRUE, TRUE, FALSE, FALSE, FALSE,
                 FALSE, FALSE, FALSE, FALSE, TRUE, TRUE),
        name = "L")
S <- pm(nrow = 2, ncol = 2,
        labels = c("s11", "s12", "s12", "s22"),
        values = c(2, 1, 1, 2),
        free = TRUE,
        name = "S")
vc <- svc(form = L \%*\% S \%*\% t(L), R = R)
X <- cbind(1, rnorm(100))
B <- pm(nrow = 6, ncol = 2,
        labels = paste0("b", 1:12),
        values = 0,
        free = TRUE,
        name = "B")
mc1 <- mc(form = B, X = X)
Y <- matrix(rnorm(100 * 4), 100, 4)
mod <- svcm(Y, L, S, vc, B, mc1)
\dontrun{
fit <- fitm(mod, se = TRUE)
}
}
