% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{ic}
\alias{ic}
\title{Constructor for intermediate calculation step object}
\usage{
ic(form, name = character())
}
\arguments{
\item{form}{An expression for the calculation.}

\item{name}{Character giving name to object.}
}
\value{
An object of class \code{ic}.
}
\description{
Creates an object of type \code{ic} used to represent intemediate calculation steps in a model.
}
