% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_print.R
\name{print.summary.svcm}
\alias{print.summary.svcm}
\title{Print function for summary.fitm}
\usage{
\method{print}{summary.svcm}(x, ...)
}
\arguments{
\item{x}{An object of type summary.svc.}

\item{...}{Not used}
}
\description{
Print function for summary.fitm objects.
}
