% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_print.R
\name{print.svc}
\alias{print.svc}
\title{Print function for svc objects.}
\usage{
\method{print}{svc}(x, ...)
}
\arguments{
\item{x}{An object of type svc.}

\item{...}{Not used.}
}
\description{
Print function for svc objects.
}
