% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcm_utilities.R
\name{summary.svcm}
\alias{summary.svcm}
\title{Summary function for svcm models.}
\usage{
\method{summary}{svcm}(object, ...)
}
\arguments{
\item{object}{An object of type \code{svcm}.}

\item{...}{Not used.}
}
\description{
Summary function for class \code{"svcm"}.
}
