% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{svc}
\alias{svc}
\title{Constructor for variance component object}
\usage{
svc(form, R = NULL)
}
\arguments{
\item{form}{An expression for the covariance model.}

\item{R}{A relationship matrix. Should be a sparse matrix from the \code{Matrix} package.}
}
\value{
An object of class \code{svc}.
}
\description{
Creates a variance component object as a function of model objects and a relationship matrix.
}
\examples{
library(svcmr)
R <- Matrix::Diagonal(100)
L <- pm(nrow = 6, ncol = 2,
        labels = paste0("l", 1:12),
        values = c(1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1),
        free = c(FALSE, TRUE, TRUE, FALSE, FALSE, FALSE,
                 FALSE, FALSE, FALSE, FALSE, TRUE, TRUE),
        name = "L")
S <- pm(nrow = 2, ncol = 2,
        labels = c("s11", "s12", "s12", "s22"),
        values = c(2, 1, 1, 2),
        free = TRUE,
        name = "S")
vc <- svc(form = L \%*\% S \%*\% t(L), R = R)
}
