\name{SVCM-package}
\alias{SVCM-package}
\alias{SVCM}
\docType{package}
\title{Spatially Varying Coefficient Model}
\description{This R package is about spatially varying coefficient model and its applications}
\details{
\tabular{ll}{
Package: \tab SVCM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-03-28\cr
License: \tab GPL-2\cr
}
The package contains one main function:

\code{Simtest} - To do simulation


}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
\examples{
pattern=6;
  n=60;
  pr=0.5;
  nf1=ef=f1=f2=1;
  Simtest(n, pattern, pr, nf1,ef, f1,f2);


}
