% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcmr_main.R
\name{svcm}
\alias{svcm}
\title{Creates a model}
\usage{
svcm(Y, ...)
}
\arguments{
\item{Y}{Matrix of data described by model.}

\item{...}{All relevant model, mean and variance components objects used to define the model.}
}
\value{
An object of type svcm.
}
\description{
Creates a new model
}
